# backend/openapi.yml
openapi: 3.1.0
info:
  version: '1'
  title: 'MyAce API Documentation'
  x-logo:
    url: 'https://myace.ai/logo.svg'
    # backgroundColor: '#FFFFFF'
    altText: 'MyAce logo'
servers:
  - url: 'https://api.myace.ai'
paths:
  /register:
    post:
      summary: Register
      description: This route creates a new application user by accepting an enterprise invitation or a new team member using the server password.
      tags:
        - User
      requestBody:
        required: true
        content:
          'application/json':
              schema:
                oneOf:
                  - $ref: "#/components/schemas/NewAppUser"
                  - $ref: "#/components/schemas/NewMyAceTeamMember"
      security: [] # no authentication
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUserWToken'
        default:
          $ref: '#/components/responses/DefaultError'
  /login:
    post:
      summary: Login
      description: This route establishes a session and returns a user.
      tags:
        - User
      requestBody:
        required: true
        content:
          'application/json':
              schema:
                type: object
                required: [email, password]
                properties:
                  email:
                    $ref: '#/components/schemas/UserEmail'
                  password:
                    $ref: '#/components/schemas/UserPassword'
      security: [] # no authentication
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUserWToken'
        default:
          $ref: '#/components/responses/DefaultError'
  /users/{user_id}:
    get:
      summary: Get individual user
      description: This route retrieves the profile of a given user. If the authentication header is submitted and `user_id` matches that of the authenticated user, sensitive information will be returned.
      tags:
        - User
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserID'
      security: [] # no authentication required
      responses:
        '200':
          description: Retrieved user
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PublicUser'
                  - $ref: '#/components/schemas/PrivateUser'
        default:
          $ref: '#/components/responses/DefaultError'
  /users/me:
    patch:
      summary: Update current user
      description: This route edits the profile of the authenticated user.
      tags:
        - User
      requestBody:
        description: All fields are optional
        required: false
        content:
          'application/json':
              schema:
                type: object
                properties:
                  username:
                    $ref: "#/components/schemas/UserUsername"
                  display_name:
                    $ref: "#/components/schemas/UserDisplayName"
                  biography:
                    $ref: "#/components/schemas/UserBio"
                  password:
                    type: object
                    description: The old/new password combination.
                    required: [old_password, new_password]
                    properties:
                      old_password:
                        $ref: '#/components/schemas/UserPassword'
                      new_password:
                        $ref: '#/components/schemas/UserPassword'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUser'
        default:
          $ref: '#/components/responses/DefaultError'

    delete:
      summary: Delete current user
      description: This route permanently deletes the account of the user who is currently logged in along with all associated information.
      tags:
        - User
      responses:
        '204':
          description: User deleted
        default:
          $ref: '#/components/responses/DefaultError'
  /usernames/{username}/check:
    get:
      summary: Check username
      description: This route checks the availability of a given username.
      tags:
        - User
      parameters:
        - name: username
          in: path
          description: The username to query. # this description overrides the schema's description
          required: true
          schema:
            $ref: '#/components/schemas/UserUsername'
      security: [] # no authentication
      responses:
        '200':
          description: Checked username
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available:
                    type: boolean
                    description: Whether or not another user has reserved this username.
        default:
          $ref: '#/components/responses/DefaultError'
...
components:
  responses:
    DefaultError:
      description: Unexpected error
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
  schemas:
    NewMyAceTeamMember:
      type: object
      required:
        - email
        - username
        - password
        - role
        - server_password
      properties:
        email:
          $ref: "#/components/schemas/UserEmail"
        username:
          $ref: "#/components/schemas/UserUsername"
        password:
          $ref: "#/components/schemas/UserPassword"
        role:
          $ref: "#/components/schemas/MyAceTeamRole"
        server_password:
          type: string
          description: The master password of the application.
          format: password
    NewAppUser:
      type: object
      required:
        - username
        - invite_code
        - password
      properties:
        username:
          $ref: "#/components/schemas/UserUsername"
        display_name:
          $ref: "#/components/schemas/UserDisplayName"
        biography:
          $ref: "#/components/schemas/UserBio"
        invite_code:
          type: string
          description: A special code that will be used to join an enterprise immediately after registration.
        password:
          $ref: "#/components/schemas/UserPassword"
    PublicUser:
      description: The public view of a user.
      type: object
      required:
        - user_id
        - username
        - display_name
        - biography
      properties:
        user_id:
          $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/UserUsername"
        display_name:
          $ref: "#/components/schemas/UserDisplayName"
        biography:
          $ref: "#/components/schemas/UserBio"
    PrivateUser:
      description: The private view of a user, which contains sensitive profile information.
      allOf:
        - $ref: '#/components/schemas/PublicUser'
        - type: object
          required:
            - email
            - created_at
            - updated_at
          properties:
            email:
              $ref: "#/components/schemas/UserEmail"
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              description: The ISO 8601 timestamp of the user's last profile edit.
              format: date-time
              nullable: true
    PrivateUserWToken:
      description: The private view of a user, which contains sensitive profile information and a refreshed session token.
      allOf:
        - $ref: "#/components/schemas/PrivateUser"
        - type: object
          required:
            - token
          properties:
            token:
              type: string
              description: A bearer token that may be used in in the Authentication header.
...
    MyAceTeamRole:
      type: string
      description: The role of a user within the MyAce development team.
      enum:
        - backend
        - frontend
        - business
    UserID:
      type: string
      description: The user's unique identifier.
      format: uuid
    UserEmail:
      type: string
      description: The user's email address.
      format: email
      pattern: ^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$
      example: johnsmith@gmail.com
    UserUsername:
      type: string
      description: A unique, user-friendly profile identifier.
      pattern: ^[a-z0-9_.]{4,16}$
      example: tennislover9
    UserPassword:
      type: string
      description: The user's password.
      format: password
      # TODO: add regex
      # pattern: ^(?=.*?[a-z])(?=.*?[A-Z])(?=.*?[0-9\-\"[\]!#$%&'()*+,./:;<=>?@^_`{|}~])[A-Za-z0-9\-\"[\]!#$%&'()*+,./:;<=>?@^_`{|}~]{6,}$
      example: Tnns31!
    UserDisplayName:
      type: string
      description: The name of a user.
      example: John Smith
    UserBio:
      type: string
      description: A personalized biography.
      example: I couldn't figure out why the ball was getting bigger. Then it hit me.
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
security:
  - bearerAuth: []